python_supabase/requirements.txt
supabase
python-dotenv
postgrest-py
httpx
